requirements.txt
pyaes==1.6.1
pyasn1==0.4.5
PySocks==1.6.8
rsa==4.0
Telethon==1.6.2
